Open the source
This project was created in 2014 as an Android game as a hobby project
on my first year of
college.
In 2016 I have uploaded this project on github to make it open source
and to show it as my
portfolio from 2014. This project has a lot of things wrong like,
almost everything is public and I
didn't know Design Patterns at time, there is a lot of room for
improvement.
This project was a great way to learn about how to create and finish a
game and opened new doors for me
to learn more and improve myself.